review_miner/requirements.txt
amqp==5.2.0
async-timeout==4.0.3
attrs==24.2.0
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coverage==7.6.8
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.15.4
greenlet==3.1.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
kombu==5.3.7
lxml==5.2.2
netifaces==0.11.0
numpy==2.1.3
packaging==24.1
pandas==2.2.3
parsel==1.9.1
playwright==1.49.0
pluggy==1.5.0
prompt_toolkit==3.0.47
Protego==0.3.1
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.0.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pytest==8.3.4
pytest-benchmark==5.1.0
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
queuelib==1.7.0
redis==5.0.8
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==75.3.0
six==1.16.0
tldextract==5.1.2
tomli==2.0.1
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
zope.interface==7.0.1
